Created folder for recording classes in project structure
Replaced IAudioRecordingWriter interface with abstract AudioLameWriterBase
Refactored both MixDownWriter and PerRadioWriter to move duplicate code and field to the abstract class
Fixed recording output directory
