+++ DCS-SR-Client/Recording/AudioManipulationHelper.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Ciribob.DCS.SimpleRadio.Standalone.Common.Helpers
+namespace Ciribob.DCS.SimpleRadio.Standalone.Client.Recording
 {
     public static class AudioManipulationHelper
     {
